Move "horizBrace" to functions/horizBrace.js.
